fixed some tests, worked on real test implementation
git-svn-id: https://pushee.googlecode.com/svn/trunk/pushee@42 1ae130c9-2b25-0426-bda3-53af0882bf06

+++ src/main/java/org/mitre/pushee/hub/web/ClientConnection.java
@@ -18,6 +18,6 @@ public interface ClientConnection {
 
 	// utility functions, to be moved to utility class
 	public boolean verifyCallback(String callback, String mode, String topic,
-			long leaseSeconds, String verifyToken);
+			int leaseSeconds, String verifyToken);
 
 }
